+++ server/src/main/scala/ch.epfl.scala.index.server/Server.scala
@@ -3,7 +3,6 @@ package server
 
 import model._
 import model.misc.UserInfo
-import model.release.SemanticVersion
 // import data.cleanup.SemanticVersionParser
 import data.elastic._
 
@@ -47,83 +46,6 @@ object Server {
         latestProjects <- api.latestProjects()
         latestReleases <- api.latestReleases()
       } yield views.html.frontpage(keywords, targets, dependencies, latestProjects, latestReleases, userInfo)
-    }
-
-    // owner: akka
-    // part2: akka (project) or akka-http (artifact)
-    def projectPage(owner: String, part2: String, version: Option[SemanticVersion], user: Option[UserInfo]) = {
-      val projectRef = Project.Reference(owner, part2)
-
-      val projectAndReleases =
-        for {
-          project <- api.project(projectRef)
-          releases <- api.releases(projectRef)
-        } yield (project, releases)
-
-      def finds[A, B](xs: List[(A, B)], fs: List[A => Boolean]): Option[(A, B)] = {
-        fs match {
-          case Nil => None
-          case f :: h =>
-            xs.find{ case (a, b) => f(a) } match {
-              case None => finds(xs, h)
-              case s    => s
-            }
-        }
-      }
-
-      def defaultRelease(project: Project, releases: List[Release]): Option[(Release, List[SemanticVersion])] = {
-        val artifacts = releases.groupBy(_.reference.artifact).toList
-        
-        def projectDefault(artifact: String): Boolean = Some(artifact) == project.defaultArtifact
-        def core(artifact: String): Boolean = artifact.endsWith("-core")
-        def projectRepository(artifact: String): Boolean = project.reference.repository == artifact
-
-        def alphabetically = artifacts.sortBy(_._1).headOption
-
-        (finds(artifacts, List(projectDefault _, core _, projectRepository _)) match {
-          case None => alphabetically
-          case x => x
-        }).flatMap{ case (_, releases) =>
-          val sortedByVersion = releases.sortBy(_.reference.version)(Descending)
-
-          // select last non preRelease release if possible (ex: 1.1.0 over 1.2.0-RC1)
-          val lastNonPreRelease =
-            if(sortedByVersion.exists(_.reference.version.preRelease.isEmpty)){
-              sortedByVersion.filter(_.reference.version.preRelease.isEmpty)
-            } else {
-              sortedByVersion
-            }
-
-          // select latest stable target if possible (ex: scala 2.11 over 2.12 and scala-js)
-          val latestStableTarget =
-            (
-              if(lastNonPreRelease.exists(_.reference.target.scalaJsVersion.isEmpty)) {
-                lastNonPreRelease.filter(_.reference.target.scalaJsVersion.isEmpty)
-              } else lastNonPreRelease
-            ).sortBy(_.reference.target.scalaVersion).headOption
-
-          latestStableTarget.headOption.map(r =>
-            (r, sortedByVersion.map(_.reference.version).distinct)
-          )
-        }
-      }
-
-      projectAndReleases.map{ case (p, releases) =>
-        p.flatMap(project =>
-          defaultRelease(project, releases).map{ case (selectedRelease, versions) =>
-            val artifacts = releases.groupBy(_.reference.artifactReference).keys.toList.map(_.artifact)
-
-            (OK, views.project.html.project(
-              project,
-              artifacts,
-              versions,
-              selectedRelease,
-              releases.size,
-              user
-            ))
-          }
-        ).getOrElse((NotFound, views.html.notfound(user)))
-      }
     }
 
     val route = {
@@ -210,27 +132,37 @@ object Server {
             }
           }
         } ~
-        path(Segment / Segment) { (owner, part2) =>
+        path(Segment / Segment) { (owner, repo) =>
           optionalSession(refreshable, usingCookies) { userState =>
-            complete(projectPage(owner, part2, version = None, userState.map(_.user)))
+            val projectRef = Project.Reference(owner, repo)
+            val user = userState.map(_.user)
+
+            complete(api.projectPage(projectRef).map(
+              _.map{ case (project, artifacts, versions, release, releaseCount) =>
+                (OK, views.project.html.project(
+                  project,
+                  artifacts,
+                  versions,
+                  release,
+                  releaseCount,
+                  user
+                ))
+              }.getOrElse((NotFound, views.html.notfound(user))))
+            )
           }
         } ~
-        // path(Segment / Segment / Segment) { (owner, artifactName, version) =>
-        //   optionalSession(refreshable, usingCookies) { userState =>
-        //     val reference = Artifact.Reference(owner, artifactName)
-        //     complete(artifactPage(reference, SemanticVersionParser(version), userState.map(_.user)))
-        //   }
-        // } ~
-        // path("edit" / Segment / Segment) { (owner, artifactName) =>
-        //   optionalSession(refreshable, usingCookies) { userState =>
-        //     val reference = Artifact.Reference(owner, artifactName)
-        //     complete(
-        //       api.projectPage(reference).map(project =>
-        //         project.map(p => views.html.editproject(p, reference, version = None, userState.map(_.user)))
-        //       )
-        //     )
-        //   }
-        // } ~
+        path(Segment / Segment / Segment) { (owner, artifact, version) =>
+          optionalSession(refreshable, usingCookies) { userState =>
+            // val reference = Artifact.Reference(owner, artifactName)
+            // complete(artifactPage(reference, SemanticVersionParser(version), userState.map(_.user)))
+            complete("OK")
+          }
+        } ~
+        path("edit" / Segment / Segment) { (owner, artifactName) =>
+          optionalSession(refreshable, usingCookies) { userState =>
+            complete("OK")
+          }
+        } ~
         pathSingleSlash {
           optionalSession(refreshable, usingCookies) { userState =>
             complete(frontPage(userState.map(_.user)))
